+++ postgres/tests/sync.rs
@@ -2,10 +2,9 @@ use core::future::IntoFuture;
 
 use xitca_postgres::{
     error::{DbError, RuntimeError, SqlState},
-    pipeline::Pipeline,
     statement::Statement,
     types::Type,
-    Client, Execute, ExecuteBlocking, Postgres,
+    Client, ExecuteBlocking, Postgres,
 };
 
 fn connect() -> Client {
@@ -88,33 +87,4 @@ async fn cancel_query_blocking_in_tokio() {
 
     let e = e.downcast_ref::<RuntimeError>().unwrap();
     assert_eq!(e, &RuntimeError::RequireNoTokio);
-}
-
-#[test]
-fn pipeline_blocking() {
-    let cli = connect();
-
-    "CREATE TEMPORARY TABLE foo (name TEXT, age INT);"
-        .execute_blocking(&cli)
-        .unwrap();
-
-    Statement::unnamed(
-        "INSERT INTO foo (name, age) VALUES ($1, $2), ($3, $4), ($5, $6);",
-        &[Type::TEXT, Type::INT4, Type::TEXT, Type::INT4, Type::TEXT, Type::INT4],
-    )
-    .bind_dyn(&[&"alice", &20i32, &"bob", &30i32, &"charlie", &40i32])
-    .execute_blocking(&cli)
-    .unwrap();
-
-    let stmt = Statement::named("UPDATE foo SET age = 30 WHERE name = $1", &[])
-        .execute_blocking(&cli)
-        .unwrap();
-
-    let mut pipe = Pipeline::new();
-
-    stmt.bind(["alice"]).query(&mut pipe).unwrap();
-    stmt.bind(["bob"]).query(&mut pipe).unwrap();
-
-    let rows_affected = pipe.execute_blocking(&cli).unwrap();
-    assert_eq!(rows_affected, 2);
 }